requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
bcrypt==4.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.8
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.1.0
mysql==0.0.3
mysql-connector-python==9.2.0
mysqlclient==2.2.7
passlib==1.7.4
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
python-http-client==3.3.7
requests==2.32.3
sendgrid==6.11.0
sniffio==1.3.1
SQLAlchemy==2.0.38
SQLAlchemy-Utils==0.41.2
starkbank-ecdsa==2.2.0
starlette==0.45.3
twilio==9.4.5
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
